scripts/requirements.txt
python-frontmatter
Pillow
pyzbar
requests

